requirements.txt
numpy==1.26.4
opencv-python==4.8.1.78
streamlit==1.28.2
pandas==2.1.4
psycopg[binary]==3.1.13
pillow==10.0.1
